+++ packages/nano-ui/test/integration/Main.hs
@@ -37,6 +37,7 @@ import Cases.CustomWidget qualified
 import Cases.WidgetIdIndex qualified
 import Cases.Paths qualified
 import Cases.LayoutFlow qualified
+import Cases.Images qualified
 import NanoUI.Testing.Runner (runTests)
 
 main :: IO ()
@@ -79,4 +80,5 @@ main =
     , Cases.WidgetIdIndex.tests
     , Cases.Paths.tests
     , Cases.LayoutFlow.tests
+    , Cases.Images.tests
     ]